Move location of errors.go from libeat to filebeat
