+++ filebeat/input/errors.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package common
+package input
 
 import (
 	"fmt"
