+++ filebeat/input/log/input.go
@@ -175,7 +175,7 @@ func (p *Input) loadStates(states []file.State) error {
 
 			// In case a input is tried to be started with an unfinished state matching the glob pattern
 			if !state.Finished {
-				return &common.ErrInputNotFinished{State: state.String()}
+				return &input.ErrInputNotFinished{State: state.String()}
 			}
 
 			// Convert state to current identifier if different
