+++ filebeat/input/runnerfactory.go
@@ -59,7 +59,7 @@ func (r *RunnerFactory) Create(
 
 func (r *RunnerFactory) CheckConfig(cfg *common.Config) error {
 	_, err := r.Create(pipeline.NewNilPipeline(), cfg)
-	if _, ok := err.(*common.ErrInputNotFinished); ok {
+	if _, ok := err.(*ErrInputNotFinished); ok {
 		// error is related to state, and hence config can be considered valid
 		return nil
 	}
